added music commands (from reconlx) and made all commands use embeds.

+++ commands/moderation/clearwarns.js
@@ -1,4 +1,5 @@
 /* eslint-disable no-empty */
+const { MessageEmbed } = require('discord.js');
 const db = require('../../models/warndb')
 
 module.exports = {
@@ -12,13 +13,36 @@ module.exports = {
         }, async (err, data) => {
             if (err) throw err;
             if (data) {
+
+                if (!message.member.permissions.has("MANAGE_MESSAGES")) {
+
+                    const noPerms = new MessageEmbed()
+        
+                        .setTitle('<:ulobError:894937662518091776> You cannot use this command!')
+                        .setColor('#F04947')
+                        .setTimestamp();
+        
+                    return message.channel.send({ embeds: [noPerms] });
+                }
+
                 await db.findOneAndDelete({
                     user: user.user.id,
                     guild: message.guild.id
                 })
-                message.channel.send(`Cleared all the warnings for ${user.displayName}`)
+                const clearReturn = new MessageEmbed()
+
+                .setTitle(`<:ulobSuccess:894937662497128488> Cleared all warnings for ${user.displayName}.`)
+                .setColor('#43B581')
+                .setTimestamp()
+
+                message.channel.send({ embeds: [clearReturn] })
             } else {
-                message.channel.send('This user does not have any warns in this server!')
+                const noWarns = new MessageEmbed()
+                .setTitle('<:ulobError:894937662518091776> This user does not have any warnings in this server!')
+                .setColor('#F04947')
+                .setTimestamp();
+
+                message.channel.send({ embeds: [noWarns] })
             }
         }) 
     }
